+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Clicking on the many of the indicators in the bottom status bar takes you to the settings view where you can change the related settings.
 - Added accordions to the terminal right-hand drawer, and added more controls such as quick port settings.
-- The terminal right-hand drawer now remembers if it was open or closed. This info is also saved in each profile. 
+- The terminal right-hand drawer now remembers if it was open or closed. This info is also saved in each profile.
+- Added to option to send a break signal on Enter key press.
 
 ### Changed